+++ packages/v1/runtime/CHANGELOG.md
@@ -1,5 +1,65 @@
 # @copilotkit/runtime
 
+## 1.52.1
+
+### Patch Changes
+
+- @copilotkit/shared@1.52.1
+- @copilotkitnext/agent@1.52.1
+- @copilotkitnext/runtime@1.52.1
+
+## 1.52.1-next.1
+
+### Patch Changes
+
+- @copilotkit/shared@1.52.1-next.1
+- @copilotkitnext/agent@1.52.1-next.1
+- @copilotkitnext/runtime@1.52.1-next.1
+
+## 1.52.1-next.0
+
+### Patch Changes
+
+- @copilotkit/shared@1.52.1-next.0
+- @copilotkitnext/agent@1.52.1-next.0
+- @copilotkitnext/runtime@1.52.1-next.0
+
+## 1.52.0
+
+### Minor Changes
+
+- 6dd6c84: Adding new v2 features:
+  - useComponent
+  - useRenderTool
+  - useDefaultRenderTool
+
+  Also, fixing issues with styles not being
+  properly scoped for tailwind.
+
+### Patch Changes
+
+- ef0f539: Add reasoning support and default components for reasoning messages
+- Updated dependencies [ef0f539]
+  - @copilotkitnext/runtime@1.52.0
+  - @copilotkit/shared@1.52.0
+  - @copilotkitnext/agent@1.52.0
+
+## 1.52.0-next.8
+
+### Patch Changes
+
+- @copilotkit/shared@1.52.0-next.8
+- @copilotkitnext/agent@1.52.0-next.8
+- @copilotkitnext/runtime@1.52.0-next.8
+
+## 1.52.0-next.7
+
+### Patch Changes
+
+- @copilotkit/shared@1.52.0-next.7
+- @copilotkitnext/agent@1.52.0-next.7
+- @copilotkitnext/runtime@1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ### Patch Changes