+++ packages/v2/voice/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @copilotkit/voice
 
+## 1.52.1
+
+### Patch Changes
+
+- @copilotkitnext/runtime@1.52.1
+
+## 1.52.1-next.1
+
+### Patch Changes
+
+- @copilotkitnext/runtime@1.52.1-next.1
+
+## 1.52.1-next.0
+
+### Patch Changes
+
+- @copilotkitnext/runtime@1.52.1-next.0
+
+## 1.52.0
+
+### Patch Changes
+
+- Updated dependencies [ef0f539]
+  - @copilotkitnext/runtime@1.52.0
+
+## 1.52.0-next.8
+
+### Patch Changes
+
+- @copilotkitnext/runtime@1.52.0-next.8
+
+## 1.52.0-next.7
+
+### Patch Changes
+
+- @copilotkitnext/runtime@1.52.0-next.7
+
 ## 1.52.0-next.6
 
 ### Patch Changes